few renames and code organization
