+++ classic.go
@@ -1,22 +1,14 @@
 package fluentbitconfig
 
 import (
-	"bytes"
-	"encoding"
-	"encoding/json"
 	"fmt"
 	"regexp"
-	"strconv"
 	"strings"
 	"text/tabwriter"
 	"unicode/utf8"
 
 	"github.com/calyptia/go-fluentbit-config/v2/property"
 )
-
-// reSpaces matches more than one consecutive spaces.
-// This is used to split keys from values from the classic config.
-var reSpaces = regexp.MustCompile(`\s+`)
 
 func (c *Config) UnmarshalClassic(text []byte) error {
 	lines := strings.Split(string(text), "\n")
@@ -216,6 +208,10 @@ func (c Config) MarshalClassic() ([]byte, error) {
 	return []byte(sb.String()), nil
 }
 
+// reSpaces matches more than one consecutive spaces.
+// This is used to split keys from values from the classic config.
+var reSpaces = regexp.MustCompile(`\s+`)
+
 func spaceCut(s string) (before, after string, found bool) {
 	parts := reSpaces.Split(s, 2)
 	if len(parts) != 2 {
@@ -223,97 +219,4 @@ func spaceCut(s string) (before, after string, found bool) {
 	}
 
 	return parts[0], parts[1], true
-}
-
-// anyFromString used to convert property values from the classic config
-// to any type.
-func anyFromString(s string) any {
-	// not using strconv since the boolean parser is not very strict
-	// and allows: 1, t, T, 0, f, F.
-	if strings.EqualFold(s, "true") {
-		return true
-	}
-
-	if strings.EqualFold(s, "false") {
-		return false
-	}
-
-	if v, err := strconv.ParseInt(s, 10, 64); err == nil {
-		return v
-	}
-
-	if v, err := strconv.ParseFloat(s, 64); err == nil {
-		return v
-	}
-
-	if u, err := strconv.Unquote(s); err == nil {
-		return u
-	}
-
-	return s
-}
-
-// stringFromAny -
-// TODO: Handle more data types.
-func stringFromAny(v any) string {
-	switch t := v.(type) {
-	case encoding.TextMarshaler:
-		if b, err := t.MarshalText(); err == nil {
-			return stringFromAny(string(b))
-		}
-	case fmt.Stringer:
-		return stringFromAny(t.String())
-	case json.Marshaler:
-		if b, err := t.MarshalJSON(); err == nil {
-			return stringFromAny(string(b))
-		}
-	case map[string]any:
-		var buff bytes.Buffer
-		enc := json.NewEncoder(&buff)
-		enc.SetEscapeHTML(false)
-		if err := enc.Encode(t); err == nil {
-			return buff.String()
-		}
-	case float32:
-		if isFloatInt(t) {
-			return strconv.FormatInt(int64(t), 10)
-		}
-		return fmtFloat(t)
-	case float64:
-		if isFloatInt(t) {
-			return strconv.FormatInt(int64(t), 10)
-		}
-		return fmtFloat(t)
-	case string:
-		if strings.Contains(t, "\n") {
-			return fmt.Sprintf("%q", t)
-		}
-
-		if t == "" {
-			return `""`
-		}
-
-		return t
-	}
-
-	return stringFromAny(fmt.Sprintf("%v", v))
-}
-
-// isFloatInt reports whether a float is an integer number
-// with no fractional part.
-func isFloatInt[F float32 | float64](f F) bool {
-	switch t := any(f).(type) {
-	case float32:
-		return t == float32(int32(f))
-	case float64:
-		return t == float64(int64(f))
-	}
-	return false
-}
-
-func fmtFloat[F float32 | float64](f F) string {
-	s := fmt.Sprintf("%f", f)
-	s = strings.TrimRight(s, "0")
-	s = strings.TrimRight(s, ".")
-	return s
 }
